lifter、storage処理追加 上記に伴いnews_documentsケース修正

+++ app/Libs/DeleteFile.php
@@ -12,8 +12,13 @@ class DeleteFile
      */
     public static function deleteFilePath($path, $fileNames)
     {
-        foreach ($fileNames as $fileName) {
-            Storage::disk('public')->delete($path . $fileName['document_path']);
+        if ($path == '/lifter-images/') {
+            Storage::disk('public')->delete($path . $fileNames['image_path']);
+        }
+        if ($path == '/news-documents/') {
+            foreach ($fileNames as $fileName) {
+                Storage::disk('public')->delete($path . $fileName['document_path']);
+            }
         }
     }
 }
